+++ lib/mineral_contract.dart
@@ -6,3 +6,6 @@ export 'src/contracts/event_service_contract.dart';
 export 'src/contracts/command_service_contract.dart';
 export 'src/contracts/shared_state_service_contract.dart';
 export 'src/contracts/context_menu_service_contract.dart';
+export 'src/contracts/environment_service_contract.dart';
+export 'src/contracts/console_service_contract.dart';
+export 'src/contracts/cli_service_contract.dart';