+++ lib/src/contracts/shared_state_service_contract.dart
@@ -1,3 +1,7 @@
+abstract class MineralStateContract<T> {}
+
 abstract class SharedStateServiceContract {
   T use<T> ();
+
+  void register (List<MineralStateContract> mineralStates);
 }