+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+## 1.4.0
+- Add some contracts
+
+## 1.3.0
+- Add Console getter support
+
+## 1.2.0
+- Add Environment getter support
+
+## 1.1.0
+- Improve bad or missing mineral component contracts
+
 ## 1.0.0
 - Initial version.